Remove pagination plugin
