+++ common/src/main/java/com/earphone/common/constant/Constants.java
@@ -7,10 +7,4 @@ package com.earphone.common.constant;
  */
 public class Constants {
     public static final String CODE = "code";
-    /*
-     * 分页常量
-     */
-    public static final String TOTAL = "total";
-    public static final String PAGE_NO = "pageNo";
-    public static final String PAGE_SIZE = "pageSize";
 }
