+++ web/.gitignore
@@ -1,0 +1,8 @@
+*.idea
+*.iml
+/.settings
+/.classpath
+/.mymetadata
+/.project
+/target
+/src/main/webapp/WEB-INF/classes
