+++ web/src/main/java/com/earphone/clone/config/TransactionConfig.java
@@ -1,10 +1,5 @@
 package com.earphone.clone.config;
 
-import com.earphone.common.plugin.pagination.CacheInterceptor;
-import com.earphone.common.plugin.pagination.MybatisInterceptor;
-import com.earphone.common.plugin.pagination.PaginationInterceptor;
-import com.earphone.common.plugin.pagination.ResultSetInterceptor;
-import org.apache.ibatis.plugin.Interceptor;
 import org.mybatis.spring.SqlSessionFactoryBean;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.context.ApplicationContext;
@@ -45,23 +40,7 @@ public class TransactionConfig {
         factoryBean.setDataSource(dataSource);
         factoryBean.setConfigLocation(new ClassPathResource("mybatis-config.xml"));
         factoryBean.setMapperLocations(context.getResources("classpath*:/com/earphone/clone/module/**/*.xml"));
-        factoryBean.setPlugins(new Interceptor[]{cacheInterceptor(), paginationInterceptor(), resultSetInterceptor()});
         return factoryBean;
-    }
-
-    @Bean
-    public MybatisInterceptor cacheInterceptor() {
-        return new CacheInterceptor();
-    }
-
-    @Bean
-    public MybatisInterceptor paginationInterceptor() {
-        return new PaginationInterceptor();
-    }
-
-    @Bean
-    public MybatisInterceptor resultSetInterceptor() {
-        return new ResultSetInterceptor();
     }
 
     @Bean
