+++ web/src/main/java/com/earphone/clone/module/constant/service/ConstantService.java
@@ -17,8 +17,4 @@ public class ConstantService {
     public Constant findByKey(String key) {
         return constantMapper.findByKey(key);
     }
-
-    public Constant findNoCacheByKey(String key) {
-        return constantMapper.findNoCacheByKey(key);
-    }
 }
